requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Development
watchdog==3.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2
